+++ src/main/java/uz/pdp/market/dto/orderList/OrderListDto.java
@@ -18,7 +18,7 @@ public class OrderListDto extends GenericDto {
 
     private String description;
 
-    private int amount;
+    private double amount;
 
     private int inAdvance;
 
@@ -34,7 +34,7 @@ public class OrderListDto extends GenericDto {
 
 
     @Builder(builderMethodName = "childBuilder")
-    public OrderListDto(Long id, String categoryId, String name, String description, int amount, int inAdvance, String currencyId, LocalDateTime registeredDate, LocalDateTime completeDate, String clientPhone, String madeById) {
+    public OrderListDto(Long id, String categoryId, String name, String description, double amount, int inAdvance, String currencyId, LocalDateTime registeredDate, LocalDateTime completeDate, String clientPhone, String madeById) {
         super(id);
         this.categoryId = categoryId;
         this.name = name;
